Added new post ml optimization. helped, but not better than alternative approach

+++ run_ae_sims_ml.py
@@ -17,14 +17,15 @@ from signals import *
 from frequencyestimator import *
 import time
 # import multiprocessing
-import pickle
 import argparse
 import pathlib
 import torch
+import pickle
 torch.set_num_threads(1)
 torch.multiprocessing.set_start_method('fork', force=True) # 'spawn' or 'fork'
 from scipy.stats import binom
 from ml_optsigns import SignModel
+import itertools
 
 from numba.core.errors import NumbaDeprecationWarning, NumbaPendingDeprecationWarning
 import warnings
@@ -32,110 +33,182 @@ import warnings
 warnings.simplefilter('ignore', category=NumbaDeprecationWarning)
 warnings.simplefilter('ignore', category=NumbaPendingDeprecationWarning)
 
+def objective_function_ll(lp, cos_signal, abs_sin, ula_signal, esprit):
+    signal = cos_signal + 1.0j * lp * abs_sin
+    R = ula_signal.get_cov_matrix_toeplitz(signal)
+    theta_est, _ = esprit.estimate_theta_toeplitz(R)
 
-def run(theta, n_samples, ula_signal, espirit, sign_model, num_median, eta=0.0, i=0):
+    theta_est = apply_correction(ula_signal, ula_signal.measurements, theta_est, theta_est)
+
+    # print(f'2*theta_est: {2*theta_est}')
+    p_same = np.cos((2 * ula_signal.depths + 1) * (theta_est)) ** 2
+
+    obj = -np.sum(
+        np.log(
+            [1e-75 + binom.pmf(ula_signal.n_samples[kk] * ula_signal.measurements[kk], ula_signal.n_samples[kk], p_same[kk]) for kk
+             in
+             range(len(ula_signal.n_samples))]))
+
+
+    # eigs = np.abs(esprit.eigs)[:2]
+    # obj = eigs[1] - eigs[0]
+
+    return obj
+
+def minimize_obj(all_signs, cos_signal, abs_sin, ula_signal, esprit, disp):
+    """
+    Find the sign variation that minimizes the objective function.
+
+    Args:
+        all_signs (list): A list of all possible sign variations to consider.
+        cos_signal (numpy.ndarray): The real part of the estimated signal.
+        abs_sin (numpy.ndarray): The absolute value of the imaginary part of the estimated signal.
+        ula_signal (ULASignal): An instance of the ULASignal class, containing information about the signal.
+        esprit (ESPRIT): An instance of the ESPRIT class, used for estimating the signal parameters.
+        disp (bool): If True, prints the current objective and the current best sign variation.
+
+    Returns:
+        numpy.ndarray: The sign variation that minimizes the objective function.
+    """
+    x_star = all_signs[0]
+    obj = objective_function_ll(np.array(all_signs[0]), cos_signal, abs_sin, ula_signal, esprit)
+    if disp:
+        print('original objective: ', objective_function_ll(np.array(all_signs[0]), cos_signal, abs_sin, ula_signal, esprit))
+
+    for x in all_signs:
+        curr_obj = objective_function_ll(np.array(x), cos_signal, abs_sin, ula_signal, esprit)
+        # print(f'current objective: {curr_obj}')
+        # print(f'current best signs: {x}')
+        if curr_obj < obj:
+            if disp:
+                print(f'new objective: {curr_obj}')
+                print(f'new best signs: {x}')
+            obj = curr_obj
+            x_star = np.array(x)
+
+    return x_star
+
+
+
+
+def apply_correction(ula_signal, measurements, theta_est, theta):
+    theta_est = np.abs(theta_est)
+    p_exact = np.cos((2 * ula_signal.depths + 1) * (theta)) ** 2
+    p_neg = np.cos((2 * ula_signal.depths + 1) * (-theta)) ** 2
+    p_o2 = np.cos((2 * ula_signal.depths + 1) * (theta_est / 2.0)) ** 2
+    p_o4 = np.cos((2 * ula_signal.depths + 1) * (theta_est / 4.0)) ** 2
+    p_same = np.cos((2 * ula_signal.depths + 1) * (theta_est)) ** 2
+    p_s2 = np.cos((2 * ula_signal.depths + 1) * (np.pi / 2 - theta_est)) ** 2
+    p_s4 = np.cos((2 * ula_signal.depths + 1) * (np.pi / 4 - theta_est)) ** 2
+    p_s2_o2 = np.cos((2 * ula_signal.depths + 1) * (np.pi / 2 - theta_est / 2)) ** 2
+    p_s4_o2 = np.cos((2 * ula_signal.depths + 1) * (np.pi / 4 - theta_est / 2)) ** 2
+
+    l_exact = np.sum(
+        np.log([1e-75 + binom.pmf(ula_signal.n_samples[kk] * measurements[kk], ula_signal.n_samples[kk],
+                                  p_exact[kk]) for kk in
+                range(len(ula_signal.n_samples))]))
+    l_neg = np.sum(
+        np.log([1e-75 + binom.pmf(ula_signal.n_samples[kk] * measurements[kk], ula_signal.n_samples[kk],
+                                  p_neg[kk]) for kk in
+                range(len(ula_signal.n_samples))]))
+    l_o2 = np.sum(
+        np.log(
+            [1e-75 + binom.pmf(ula_signal.n_samples[kk] * measurements[kk], ula_signal.n_samples[kk], p_o2[kk]) for kk
+             in
+             range(len(ula_signal.n_samples))]))
+    l_o4 = np.sum(
+        np.log(
+            [1e-75 + binom.pmf(ula_signal.n_samples[kk] * measurements[kk], ula_signal.n_samples[kk], p_o4[kk]) for kk
+             in
+             range(len(ula_signal.n_samples))]))
+    l_same = np.sum(
+        np.log(
+            [1e-75 + binom.pmf(ula_signal.n_samples[kk] * measurements[kk], ula_signal.n_samples[kk], p_same[kk]) for kk
+             in
+             range(len(ula_signal.n_samples))]))
+    l_s2 = np.sum(
+        np.log(
+            [1e-75 + binom.pmf(ula_signal.n_samples[kk] * measurements[kk], ula_signal.n_samples[kk], p_s2[kk]) for kk
+             in
+             range(len(ula_signal.n_samples))]))
+    l_s4 = np.sum(
+        np.log(
+            [1e-75 + binom.pmf(ula_signal.n_samples[kk] * measurements[kk], ula_signal.n_samples[kk], p_s4[kk]) for kk
+             in
+             range(len(ula_signal.n_samples))]))
+    l_s2_o2 = np.sum(
+        np.log(
+            [1e-75 + binom.pmf(ula_signal.n_samples[kk] * measurements[kk], ula_signal.n_samples[kk], p_s2_o2[kk]) for
+             kk in
+             range(len(ula_signal.n_samples))]))
+    l_s4_o2 = np.sum(
+        np.log(
+            [1e-75 + binom.pmf(ula_signal.n_samples[kk] * measurements[kk], ula_signal.n_samples[kk], p_s4_o2[kk]) for
+             kk in
+             range(len(ula_signal.n_samples))]))
+
+    which_correction = np.argmax([l_same, l_s2, l_s4, l_o2, l_o4, l_s2_o2, l_s4_o2])
+    if which_correction == 1:
+        theta_est = np.pi / 2.0 - theta_est
+    elif which_correction == 2:
+        theta_est = np.pi / 4.0 - theta_est
+    elif which_correction == 3:
+        theta_est = theta_est / 2
+    elif which_correction == 4:
+        theta_est = theta_est / 4
+    elif which_correction == 5:
+        theta_est = np.pi / 2.0 - 0.5 * theta_est
+    elif which_correction == 6:
+        theta_est = np.pi / 4.0 - 0.5 * theta_est
+    # elif which_correction == 7:
+    #     theta_est = -theta_est
+
+    # print(f'FINAL ANGLE FOUND: {theta_est, theta}')
+
+    return np.abs(theta_est)
+
+
+def run(theta, n_samples, ula_signal, espirit, sign_model, eta=0.0, i=0, thresh=1.0):
 
     torch.set_num_threads(1)
     np.random.seed(i)
-
-    median_theta = np.zeros(num_median, dtype=float)
-    # for j in range(num_median):
 
     ula_signal.estimate_signal(n_samples, theta, eta)
     X = torch.from_numpy(ula_signal.measurements).type(torch.float)
     signs = -1 + 2 * (sign_model(X) > 0.5).float().numpy()
     signs = [1.0] + list(signs)
 
-    for j in range(num_median):
-        # offset = np.random.uniform(0.0, 0.25 * np.pi)
-        offset = 0.0
-        # print(f'offset: {offset}')
-        # print(signs)
-        # offset_signal = ula_signal.add_signal_offset(ula_signal.signal, offset)
-        # signal_signs = ula_signal.update_signal_signs(offset_signal, signs)
+    sign_confidence = sign_model(X).detach().numpy()
+    # print('sign_confidence: ', sign_confidence)
+    sign_confidence_sort = np.argsort(np.abs(sign_model(X).detach().numpy()))
+    num_signs_to_switch = np.sum(np.abs(sign_confidence) < thresh)
+    signs_to_switch = sign_confidence_sort[:num_signs_to_switch]
+    # print(f'signs to switch: {signs_to_switch + 2}')
+    sign_combinations = [s for s in itertools.product([1.0, -1.0], repeat=num_signs_to_switch)]
 
-        noise_signal = ula_signal.add_signal_noise(ula_signal.signal, 0.0001)
-        signal_signs = ula_signal.update_signal_signs(noise_signal, signs)
+    all_signs = [[x for x in signs] for _ in range(len(sign_combinations))]
+    for i in range(len(sign_combinations)):
+        for j in range(num_signs_to_switch):
+            all_signs[i][signs_to_switch[j]+1] = sign_combinations[i][j]
 
-        signal = ula_signal.update_signal_signs(signal_signs, signs)
-        R = ula_signal.get_cov_matrix_toeplitz(signal)
-        theta_found, eigs = espirit.estimate_theta_toeplitz(R)
+    cos_signal = np.real(ula_signal.signal)
+    abs_sin = np.abs(np.imag(ula_signal.signal))
+    # x_star = all_signs[0]
+    x_star = minimize_obj(all_signs, cos_signal, abs_sin, ula_signal, espirit, False)
+    # print("x_star: ", x_star)
+    signal = cos_signal + 1.0j * np.array(x_star) * abs_sin
+    # signal = ula_signal.update_signal_signs(ula_signal.signs_exact, x_star)
 
-        # print(f'theta_est: {theta_found/ np.pi}')
-        # print(f'offset:    {offset / np.pi}')
+    # signal = ula_signal.update_signal_signs(ula_signal.signal, signs)
+    R = ula_signal.get_cov_matrix_toeplitz(signal)
+    # This estimates the angle using the ESPIRIT algorithm
+    theta_est, eigs = espirit.estimate_theta_toeplitz(R)
 
-        # if -np.angle(eigs) < 0:
-        #     theta_found = theta_found - offset
-        # else:
-        #     theta_found= np.pi / 2.0 - theta_found - offset
-        # print(f'theta_est: {theta_found / np.pi}')
+    theta_est = apply_correction(ula_signal, ula_signal.measurements, theta_est, theta)
 
-        p_o2 = np.cos((2 * ula_signal.depths + 1) * (theta_found / 2.0)) ** 2
-        p_o4 = np.cos((2 * ula_signal.depths + 1) * (theta_found / 4.0)) ** 2
-        p_same = np.cos((2 * ula_signal.depths + 1) * (theta_found)) ** 2
-        p_s2 = np.cos((2 * ula_signal.depths + 1) * (np.pi / 2 - theta_found)) ** 2
-        p_s4 = np.cos((2 * ula_signal.depths + 1) * (np.pi / 4 - theta_found)) ** 2
-        p_s2_o2 = np.cos((2 * ula_signal.depths + 1) * (np.pi / 2 - theta_found / 2)) ** 2
-        p_s4_o2 = np.cos((2 * ula_signal.depths + 1) * (np.pi / 4 - theta_found/ 2)) ** 2
+    error = np.abs(np.sin(theta) - np.sin(theta_est))
 
-        l_o2 = np.sum(
-            np.log(
-                [1e-75+binom.pmf(ula_signal.n_samples[kk] * ula_signal.measurements[kk], ula_signal.n_samples[kk], p_o2[kk]) for
-                 kk in
-                 range(len(ula_signal.n_samples))]))
-        l_o4 = np.sum(
-            np.log(
-                [1e-75+binom.pmf(ula_signal.n_samples[kk] * ula_signal.measurements[kk], ula_signal.n_samples[kk], p_o4[kk]) for
-                 kk in
-                 range(len(ula_signal.n_samples))]))
-        l_same = np.sum(
-            np.log(
-                [1e-75+binom.pmf(ula_signal.n_samples[kk] * ula_signal.measurements[kk], ula_signal.n_samples[kk], p_same[kk]) for
-                 kk in
-                 range(len(ula_signal.n_samples))]))
-        l_s2 = np.sum(
-            np.log(
-                [1e-75+binom.pmf(ula_signal.n_samples[kk] * ula_signal.measurements[kk], ula_signal.n_samples[kk], p_s2[kk]) for
-                 kk in
-                 range(len(ula_signal.n_samples))]))
-        l_s4 = np.sum(
-            np.log(
-                [1e-75+binom.pmf(ula_signal.n_samples[kk] * ula_signal.measurements[kk], ula_signal.n_samples[kk], p_s4[kk]) for
-                 kk in
-                 range(len(ula_signal.n_samples))]))
-        l_s2_o2 = np.sum(
-            np.log([1e-75+binom.pmf(ula_signal.n_samples[kk] * ula_signal.measurements[kk], ula_signal.n_samples[kk], p_s2_o2[kk])
-                    for kk in
-                    range(len(ula_signal.n_samples))]))
-        l_s4_o2 = np.sum(
-            np.log([1e-75+binom.pmf(ula_signal.n_samples[kk] * ula_signal.measurements[kk], ula_signal.n_samples[kk], p_s4_o2[kk])
-                    for kk in
-                    range(len(ula_signal.n_samples))]))
-
-
-        # which_correction = np.argmin([obj_same, obj_s2, obj_s4, obj_o2, obj_s2_o2, obj_s4_o2])
-        which_correction = np.argmax([l_same, l_s2, l_s4, l_o2, l_o4, l_s2_o2, l_s4_o2])
-        if which_correction == 1:
-            theta_found = np.pi / 2.0 - theta_found
-        elif which_correction == 2:
-            theta_found = np.pi / 4.0 - theta_found
-        elif which_correction == 3:
-            theta_found =  0.5 * theta_found
-        elif which_correction == 4:
-            theta_found =  0.25 * theta_found
-        elif which_correction == 5:
-            theta_found = np.pi / 2.0 - 0.5 * theta_found
-        elif which_correction == 6:
-            theta_found = np.pi / 4.0 - 0.5 * theta_found
-
-        median_theta[j] = theta_found
-
-    # print(median_theta)
-    theta_found = np.median(median_theta)
-
-    error = np.abs(np.sin(theta) - np.sin(theta_found))
-
-    return error, theta_found
+    return error, theta_est
 
 def run_process(theta, n_samples, ula_signal, espirit, sign_model, return_dict, eta=0.0, i=0):
 
@@ -241,9 +314,9 @@ if __name__ == "__main__":
     parser.add_argument('--C', type=float,
                         help="This is a free parameter that determines how many shots to take at each step. (default=3)",
                         default=3)
-    parser.add_argument('--num_median', type=int,
-                        help="Number of angle estimates to use to estimate median (default=3)",
-                        default=3)
+    parser.add_argument('--thresh', type=float,
+                        help="This is a free parameter that determines the threshold to optimize signs after ML inference. (default=1.0)",
+                        default=1.0)
     args = parser.parse_args()
 
     print('Command Line Arguments')
@@ -287,6 +360,7 @@ if __name__ == "__main__":
             espirit = ESPIRIT()
 
             narray = [2] * (2 * r + 2)
+            # narray[1] = 3
 
             ula_signal = TwoqULASignal(M=narray, C=args.C)
             print(ula_signal.depths)
@@ -301,9 +375,9 @@ if __name__ == "__main__":
             for x in narray:
                 file_subscript += f'{x}'
             filename = f'ml_models/sign_model_{file_subscript}_C{args.C:0.2f}.pt'
-            sign_model.load_state_dict(torch.load(filename, weights_only=True))
+            sign_model.load_state_dict(torch.load(filename, weights_only=False, map_location='cpu'))
             sign_model.eval()
-            sign_model.share_memory()
+            #sign_model.share_memory()
 
             arrays.append(narray)
             print(f'Array parameters: {narray}')
@@ -323,13 +397,22 @@ if __name__ == "__main__":
             num_queries[r] = np.sum(np.array(ula_signal.depths) * np.array(n_samples)) + n_samples[0]/2
             max_single_query[r] = np.max(ula_signal.depths)
 
-            pool = torch.multiprocessing.Pool(num_threads)
             start = time.time()
 
-            processes = [pool.apply_async(run, args=(theta, n_samples, ula_signal, espirit, sign_model, args.num_median, args.eta, i)) for i in
-                         range(num_mc)]
-            sims = [p.get() for p in processes]
-            pool.terminate()
+            if num_threads==1:
+                sims=[]
+                for i in range(num_mc):
+                    sim = run(theta, n_samples, ula_signal, espirit, sign_model, args.eta, i, args.thresh)
+                    sims.append(sim)
+            else:
+
+                pool = torch.multiprocessing.Pool(num_threads)
+
+                processes = [pool.apply_async(run, args=(theta, n_samples, ula_signal, espirit, sign_model, args.eta, i, args.thresh)) for i in
+                            range(num_mc)]
+                sims = [p.get() for p in processes]
+                pool.terminate()
+
             for k in range(num_mc):
                 errors[r, k], thetas[r, k] = sims[k]
 
@@ -367,3 +450,5 @@ if __name__ == "__main__":
             with open(filename, 'wb') as handle:
                 pickle.dump((errors, thetas, num_queries, max_single_query, arrays, num_lengths, num_mc),
                             handle, protocol=pickle.HIGHEST_PROTOCOL)
+                # pickle.dump((errors, thetas, num_queries, max_single_query, arrays, num_lengths, num_mc),
+                #             handle)
